Parse the OredictMaterials; Import GUI pt 4.

+++ src/main/java/com/canisartorus/prospectorjournal/lib/GeoTag.java
@@ -1,15 +1,15 @@
 package com.canisartorus.prospectorjournal.lib;
 
 public class GeoTag {
-	public final int dim, ore;
+	public final short dim, ore;
 	public int x, z; //, cx, cz;
 	public boolean sample = true;
 	
 	public GeoTag(int ore, int dim, int x, int z, boolean flower) {
-		this.dim = dim;
+		this.dim = (short) dim;
 		this.x = x;
 		this.z = z;
-		this.ore = ore;
+		this.ore = (short) ore;
 		this.sample = flower;
 //		this.cx = x / 16;
 //		this.cz = z / 16;
